@@ -7,14 +7,22 @@
   "dependencies": {
     "babel-core": "^5.8.21",
     "babel-plugin-object-assign": "^1.2.1",
+    "babel-preset-es2015": "^6.22.0",
     "babelify": "^6.1.3",
-    "bootstrap": "^3.3.5",
+    "bootstrap": "^3.3.7",
     "browserify-middleware": "^7.0.0",
+    "csv": "^1.1.1",
     "express": "^4.13.3",
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.1.1",
     "less-middleware": "^2.0.1",
+    "lodash": "^4.17.4",
+    "moment-timezone": "^0.5.11",
     "nunjucks": "^1.3.4",
-    "react": "^0.13.3",
-    "react-router": "^0.13.3"
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-router": "^3.0.2",
+    "whatwg-fetch": "^2.0.2"
   },
   "scripts": {
     "start": "node server.js",
